% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_to_polytope.R
\name{file_to_polytope}
\alias{file_to_polytope}
\title{function to get an ine or an ext file and returns the corresponding polytope}
\usage{
file_to_polytope(path, zonotope = FALSE)
}
\arguments{
\item{path}{A string that containes the path to an ine or a ext file. The ine file desrcibes a H-polytope and ext file describes a V-polytope or a zonotope.}

\item{zonotope}{A boolean parameter. It has to be TRUE when the path leads to an .ext file that describes a zonotope.}
}
\value{
A polytope class. If the path corresponds to an ine file then the return value represents a H-polytope. If it corresponds to an ext file the return value represents a V-polytope (default choice) or a zonotope if the second argument is TRUE.
}
\description{
For an ".ine" file it generates the corresponding H-polytope. For an ".ext" file it generates the corresponding V-polytope or zonotope.
For more details on those file formats see \url{https://github.com/GeomScale/volume_approximation/blob/develop/doc/cpp_interface.md#polytope-input}.
}
